Update LAB_01_RBAC.md
Fixed a typo changing "do" to "to" in the following line:
Old - 1. In the Bash session within the Cloud Shell pane, run the following to obtain a reference -> do <- the user account of Dylan Williams
Update - 1. In the Bash session within the Cloud Shell pane, run the following to obtain a reference -> to <- the user account of Dylan Williams