Update LAB_13_Azure Monitor.md
In the PowerShell session within the Cloud Shell pane, run the following to create a new Azure virtual machine.

Attention: The New-AzVm command doesn't work in the Azure CLI version 4.24 and Microsoft is currently investigating for resolution. The workaround in this lab is to install and revert back to Az.Compute version 4.23.0, which is unaffected by this issue.